+++ app/src/main/java/com/zerodsoft/scheduleweather/calendarview/interfaces/OnEventItemClickListener.java
@@ -4,6 +4,6 @@ public interface OnEventItemClickListener
 {
     void onClicked(long start, long end);
 
-    void onClicked(int calendarId, int eventId);
+    void onClicked(int calendarId, long eventId);
 }
